telephony: Release pending network requests when releasing a request
During system startup, a network request may move from the DctController
to a network factory where it is held as a pending request until the
telephony state is fully initialized.

When handling a releaseNetwork call, we must release the request
from the pending queues of all network factories as well.

Without this fix, on boot it is possible for a network request to be
added and thne moved to the pending queue of a network factory.  Then,
an attempt to release it will be a no-op instead of actually releasing
the network.

The end result is that the network request will eventually be executed
even though it is no longer desired.

If WiFi is connected when the request is executed, then the cellular
network will get a data connection but the ConnectivityService will
deem the network uneeded and kill it.  After killing it, telephony
will automatically reconnect the network because it (incorrectly)
believes that the network is needed because it was not released
when it was supposed to be.

Change-Id: Ifcb8676aec7a048e11de5a62103eb5f480566468